+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## 3.3 Changes on 2022-04
 - Add State to WebPaymentCardAddress
 - Add 3ds2 fields to WebPaymentRequestModel
+- Add CheckCard WebPayments
 - Update Api-Version to 6.13
 
 ## 3.2 Changes on 2022-01